Split `executable` module
